requirements.txt
matplotlib==3.*
numpy==1.*
pillow==10.*
python-dotenv==1.*
requests==2.*
seaborn==0.*
Flask==3.*